requirements.txt
django-money
datetime
psycopg
django_tables2
django-htmx